requirements.txt
# Streamlit UI
streamlit>=1.33.0

# Config
python-dotenv>=1.0.0

# PDF
pymupdf>=1.24.0

# RAG / LLM
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-text-splitters>=0.3.0
faiss-cpu>=1.8.0
openai>=1.40.0

# Gemini (frontend rewrites, commentaire)
google-generativeai>=0.8.0

# Snowflake (startup_rag, test_snowflake_connection)
snowflake-connector-python>=3.10.0

# Misc
requests>=2.31.0
